+++ R/data.R
@@ -1,0 +1,10 @@
+#' Example data from ggplot2 package
+#'
+#' This ggplot2 data is great for practice use in visualizations (used in function examples for y2artisan)
+#'
+#' @docType data
+#'
+#' @format A data frame with 234 observations and 11 variables
+#'
+#' @keywords datasets
+"mpg"